refactor(multiverse): Move the RoomList widget into a separate module
