+++ labs/multiverse/src/main.rs
@@ -38,6 +38,10 @@ use tokio::{runtime::Handle, spawn, task::JoinHandle};
 use tracing::{error, warn};
 use tracing_subscriber::EnvFilter;
 
+mod room_list;
+
+use room_list::{ExtraRoomInfo, RoomList};
+
 const HEADER_BG: Color = tailwind::BLUE.c950;
 const NORMAL_ROW_COLOR: Color = tailwind::SLATE.c950;
 const ALT_ROW_COLOR: Color = tailwind::SLATE.c900;
@@ -83,149 +87,6 @@ async fn main() -> Result<()> {
     app.run(terminal).await
 }
 
-#[derive(Default)]
-struct RoomList {
-    state: ListState,
-    rooms: Arc<Mutex<Vector<room_list_service::Room>>>,
-    /// Extra information about rooms.
-    room_infos: Arc<Mutex<HashMap<OwnedRoomId, ExtraRoomInfo>>>,
-}
-
-impl RoomList {
-    /// Focus the list on the next item, wraps around if needs be.
-    ///
-    /// Returns the index only if there was a meaningful change.
-    fn next(&mut self) -> Option<usize> {
-        let num_items = self.rooms.lock().unwrap().len();
-
-        // If there's no item to select, leave early.
-        if num_items == 0 {
-            self.state.select(None);
-            return None;
-        }
-
-        // Otherwise, select the next one or wrap around.
-        let prev = self.state.selected();
-        let new = prev.map_or(0, |i| if i >= num_items - 1 { 0 } else { i + 1 });
-
-        if prev != Some(new) {
-            self.state.select(Some(new));
-            Some(new)
-        } else {
-            None
-        }
-    }
-
-    /// Focus the list on the previous item, wraps around if needs be.
-    ///
-    /// Returns the index only if there was a meaningful change.
-    fn previous(&mut self) -> Option<usize> {
-        let num_items = self.rooms.lock().unwrap().len();
-
-        // If there's no item to select, leave early.
-        if num_items == 0 {
-            self.state.select(None);
-            return None;
-        }
-
-        // Otherwise, select the previous one or wrap around.
-        let prev = self.state.selected();
-        let new = prev.map_or(0, |i| if i == 0 { num_items - 1 } else { i - 1 });
-
-        if prev != Some(new) {
-            self.state.select(Some(new));
-            Some(new)
-        } else {
-            None
-        }
-    }
-}
-
-impl Widget for &mut RoomList {
-    fn render(self, area: Rect, buf: &mut Buffer)
-    where
-        Self: Sized,
-    {
-        // We create two blocks, one is for the header (outer) and the other is for list
-        // (inner).
-        let outer_block = Block::default()
-            .borders(Borders::NONE)
-            .fg(TEXT_COLOR)
-            .bg(HEADER_BG)
-            .title("Room list")
-            .title_alignment(Alignment::Center);
-        let inner_block =
-            Block::default().borders(Borders::NONE).fg(TEXT_COLOR).bg(NORMAL_ROW_COLOR);
-
-        // We get the inner area from outer_block. We'll use this area later to render
-        // the table.
-        let outer_area = area;
-        let inner_area = outer_block.inner(outer_area);
-
-        // We can render the header in outer_area.
-        outer_block.render(outer_area, buf);
-
-        // Don't keep this lock too long by cloning the content. RAM's free these days,
-        // right?
-        let mut room_info = self.room_infos.lock().unwrap().clone();
-
-        // Iterate through all elements in the `items` and stylize them.
-        let items: Vec<ListItem<'_>> = self
-            .rooms
-            .lock()
-            .unwrap()
-            .iter()
-            .enumerate()
-            .map(|(i, room)| {
-                let bg_color = match i % 2 {
-                    0 => NORMAL_ROW_COLOR,
-                    _ => ALT_ROW_COLOR,
-                };
-
-                let line = {
-                    let room_id = room.room_id();
-                    let room_info = room_info.remove(room_id);
-
-                    let (raw, display, is_dm) = if let Some(info) = room_info {
-                        (info.raw_name, info.display_name, info.is_dm)
-                    } else {
-                        (None, None, None)
-                    };
-
-                    let dm_marker = if is_dm.unwrap_or(false) { "🤫" } else { "" };
-
-                    let room_name = if let Some(n) = display {
-                        format!("{n} ({room_id})")
-                    } else if let Some(n) = raw {
-                        format!("m.room.name:{n} ({room_id})")
-                    } else {
-                        room_id.to_string()
-                    };
-
-                    format!("#{i}{dm_marker} {}", room_name)
-                };
-
-                let line = Line::styled(line, TEXT_COLOR);
-                ListItem::new(line).bg(bg_color)
-            })
-            .collect();
-
-        // Create a List from all list items and highlight the currently selected one.
-        let items = List::new(items)
-            .block(inner_block)
-            .highlight_style(
-                Style::default()
-                    .add_modifier(Modifier::BOLD)
-                    .add_modifier(Modifier::REVERSED)
-                    .fg(SELECTED_STYLE_FG),
-            )
-            .highlight_symbol(">")
-            .highlight_spacing(HighlightSpacing::Always);
-
-        StatefulWidget::render(items, inner_area, buf, &mut self.state);
-    }
-}
-
 #[derive(Default, PartialEq)]
 enum DetailsMode {
     ReadReceipts,
@@ -239,19 +100,6 @@ struct Timeline {
     timeline: Arc<SdkTimeline>,
     items: Arc<Mutex<Vector<Arc<TimelineItem>>>>,
     task: JoinHandle<()>,
-}
-
-/// Extra room information, like its display name, etc.
-#[derive(Clone)]
-struct ExtraRoomInfo {
-    /// Content of the raw m.room.name event, if available.
-    raw_name: Option<String>,
-
-    /// Calculated display name for the room.
-    display_name: Option<String>,
-
-    /// Is the room a DM?
-    is_dm: Option<bool>,
 }
 
 struct App {
@@ -415,7 +263,7 @@ impl App {
 
         Ok(Self {
             sync_service,
-            room_list: RoomList { state: Default::default(), rooms, room_infos },
+            room_list: RoomList::new(rooms, room_infos),
             client,
             listen_task,
             last_status_message: Default::default(),
